fix: skip HCL recalc without valid time; fix _updateIntervalMs overflow (uint16->uint32)

+++ src/HCL/HCLMasterManager.h
@@ -82,8 +82,8 @@ public:
     /** Last calculated time-of-day in minutes (or 0xFFFF before first loop). */
     uint16_t getLastTimeMinutes() const { return _lastTimeMinutes; }
 
-    void setUpdateInterval(uint16_t seconds) { _updateIntervalMs = seconds * 1000; }
-    uint16_t getUpdateInterval() const { return _updateIntervalMs / 1000; }
+    void setUpdateInterval(uint16_t seconds) { _updateIntervalMs = static_cast<uint32_t>(seconds) * 1000UL; }
+    uint16_t getUpdateInterval() const { return static_cast<uint16_t>(_updateIntervalMs / 1000UL); }
 
     void setFadeDuration(uint8_t seconds) { _fadeDurationSec = seconds; }
     uint8_t getFadeDuration() const { return _fadeDurationSec; }
@@ -100,7 +100,7 @@ private:
     IMasterProvider* _provider = nullptr;
     bool _enabled;
     bool _applyBlocked;
-    uint16_t _updateIntervalMs;
+    uint32_t _updateIntervalMs;
     uint8_t _fadeDurationSec;
     uint32_t _lastUpdateMs;
     uint16_t _lastTimeMinutes;
